+++ chronograf.go
@@ -958,6 +958,7 @@ type BuildStore interface {
 // that were set on the server
 type Environment struct {
 	TelegrafSystemInterval time.Duration `json:"telegrafSystemInterval"`
+	HostPageDisabled       bool          `json:"HostPageDisabled"`
 }
 
 // KVClient defines what each kv store should be capable of.