+++ ui/src/shared/apis/env.ts
@@ -2,6 +2,7 @@ import AJAX from 'src/utils/ajax'
 
 const DEFAULT_ENVS = {
   telegrafSystemInterval: '1m',
+  hostPageDisabled: false,
 }
 
 export const getEnv = async url => {