+++ ui/src/types/env.ts
@@ -1,0 +1,4 @@
+export interface Env {
+  telegrafSystemInterval: string
+  hostPageDisabled: boolean
+}